created new folder named modules, added todo and user folders to modules. added components and store folder. moved TasksAllList and renamed it to TaskCard.
